+++ logger_noop.go
@@ -1,15 +1,15 @@
 package logur
 
-type NoopLogger struct{}
+type noopLogger struct{}
 
-// NewNoop creates a no-op logger that can be used to silence
+// NewNoopLogger creates a no-op logger that can be used to silence
 // all logging from this library. Also useful in tests.
-func NewNoop() *NoopLogger {
-	return &NoopLogger{}
+func NewNoopLogger() Logger {
+	return &noopLogger{}
 }
 
-func (*NoopLogger) Trace(msg string, fields map[string]interface{}) {}
-func (*NoopLogger) Debug(msg string, fields map[string]interface{}) {}
-func (*NoopLogger) Info(msg string, fields map[string]interface{})  {}
-func (*NoopLogger) Warn(msg string, fields map[string]interface{})  {}
-func (*NoopLogger) Error(msg string, fields map[string]interface{}) {}
+func (*noopLogger) Trace(msg string, fields map[string]interface{}) {}
+func (*noopLogger) Debug(msg string, fields map[string]interface{}) {}
+func (*noopLogger) Info(msg string, fields map[string]interface{})  {}
+func (*noopLogger) Warn(msg string, fields map[string]interface{})  {}
+func (*noopLogger) Error(msg string, fields map[string]interface{}) {}