+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Exported the log testing library so it can be used for testing in other libraries
+- Unexport noop logger
 
 
 ## [0.6.0] - 2018-12-21